requirements.txt
streamlit==1.43.2
langchain-openai==0.3.8
langchain-community>=0.0.27
docarray>=0.32.0
pypdf==5.3.1
python-dotenv==1.0.1
numpy>=1.24.0
